Make the e::buffer more efficient.
It used to be that each buffer was backed by a vector.  This had the
unfortunate consequence that if one had a pointer to a buffer, it would
actually need to follow a doubly-indirect pointer.  This fixes the
buffer to allocate a fixed size buffer contiguous with all other data
about the buffer.  The fixed size buffer is a change in API that is less
flexible than before, but should also be significantly faster.