Padding selectively zeroes the buffer.
The padding object would zero all padding that was filled in.  This was
undesirable for instance, when trying to skip a number because it was
already packed appropriately.  Not zeroing the padding is problematic as
it may leak information.  The fix is to only zero padding that extends
the size of a buffer.